fixup! Add odoo decoders
