+++ account_invoice_google_document_ai/models/__init__.py
@@ -1,3 +1,4 @@
 from . import account_move
 from . import res_company
 from . import res_config_settings
+from . import account_move_google_document_ai
