+++ account_invoice_google_document_ai/models/res_company.py
@@ -13,4 +13,12 @@ class ResCompany(models.Model):
     ocr_google_processor = fields.Char()
     ocr_google_authentication = fields.Binary()
     ocr_google_authentication_name = fields.Char()
-    ocr_google_enabled = fields.Boolean()
+    ocr_google_enabled = fields.Selection(
+        [
+            ("no_send", "Do not send invoices to google"),
+            ("send_manual", "Only send invoices to google manually"),
+            ("send_automatically", "Send invoices to google automatically"),
+        ],
+        default="no_send",
+        required=True,
+    )
