+++ account_invoice_google_document_ai/tests/test_google_document_ai.py
@@ -23,7 +23,7 @@ class TestGoogleDocumentAi(AccountTestInvoicingCommon):
         super().setUpClass(chart_template_ref=chart_template_ref)
         cls.company_data["company"].write(
             {
-                "ocr_google_enabled": True,
+                "ocr_google_enabled": "send_manual",
                 "ocr_google_location": "eu",
                 "ocr_google_processor": "processor",
                 "ocr_google_project": "project",
@@ -32,7 +32,8 @@ class TestGoogleDocumentAi(AccountTestInvoicingCommon):
             }
         )
 
-    def test_ocr_process(self):
+    def test_ocr_process_manually(self):
+        self.company_data["company"].ocr_google_enabled = "send_manual"
         move = self.init_invoice("in_invoice", self.env["res.partner"], "2023-01-01")
         self.assertFalse(move.line_ids)
         move.message_post(
@@ -70,3 +71,84 @@ class TestGoogleDocumentAi(AccountTestInvoicingCommon):
             process.return_value = resp
             move.ocr_process()
         self.assertTrue(move.invoice_line_ids)
+
+    def test_ocr_process_automatically(self):
+        self.company_data["company"].ocr_google_enabled = "send_automatically"
+        move = self.init_invoice("in_invoice", self.env["res.partner"], "2023-01-01")
+        self.assertFalse(move.line_ids)
+
+        with patch.object(
+            service_account.Credentials, "from_service_account_info"
+        ) as factory, patch.object(
+            documentai_v1.DocumentProcessorServiceClient, "process_document"
+        ) as process:
+            factory.return_value = ga_credentials.AnonymousCredentials()
+            resp = document_processor_service.ProcessResponse()
+            response = document_processor_service.ProcessResponse.pb(resp)
+            json_format.Parse(
+                tools.file_open(
+                    "result.json",
+                    subdir="addons/account_invoice_google_document_ai/tests",
+                )
+                .read()
+                .encode("UTF-8"),
+                response,
+                ignore_unknown_fields=True,
+            )
+            process.return_value = resp
+            move.message_post(
+                attachments=[
+                    (
+                        "filename.pdf",
+                        bytes(
+                            tools.file_open(
+                                "in_invoice_yourcompany_demo_1.pdf",
+                                mode="rb",
+                                subdir="addons/l10n_generic_coa/static/src/demo",
+                            ).read()
+                        ),
+                    )
+                ]
+            )
+        self.assertTrue(move.invoice_line_ids)
+
+    def test_ocr_process_automatically_create(self):
+        self.company_data["company"].ocr_google_enabled = "send_automatically"
+        attachment = self.env["ir.attachment"].create(
+            {
+                "name": "filename.pdf",
+                "datas": base64.b64encode(
+                    bytes(
+                        tools.file_open(
+                            "in_invoice_yourcompany_demo_1.pdf",
+                            mode="rb",
+                            subdir="addons/l10n_generic_coa/static/src/demo",
+                        ).read()
+                    )
+                ),
+            }
+        )
+
+        with patch.object(
+            service_account.Credentials, "from_service_account_info"
+        ) as factory, patch.object(
+            documentai_v1.DocumentProcessorServiceClient, "process_document"
+        ) as process:
+            factory.return_value = ga_credentials.AnonymousCredentials()
+            resp = document_processor_service.ProcessResponse()
+            response = document_processor_service.ProcessResponse.pb(resp)
+            json_format.Parse(
+                tools.file_open(
+                    "result.json",
+                    subdir="addons/account_invoice_google_document_ai/tests",
+                )
+                .read()
+                .encode("UTF-8"),
+                response,
+                ignore_unknown_fields=True,
+            )
+            process.return_value = resp
+            self.env["account.journal"].with_context(
+                default_move_type="in_invoice"
+            ).create_invoice_from_attachment(attachment.ids)
+        # self.assertTrue(move.invoice_line_ids)
