+++ account_invoice_google_document_ai/views/res_config_settings.xml
@@ -15,9 +15,7 @@
                     class="col-12 col-lg-6 o_setting_box"
                     id="account_ocr_google_settings"
                 >
-                    <div class="o_setting_left_pane">
-                        <field name="ocr_google_enabled" />
-                    </div>
+                    <div class="o_setting_left_pane" />
                     <div
                         class="o_setting_right_pane"
                         id="account_ocr_google_settings_right"
@@ -32,19 +30,27 @@
                         />
                         <div
                             class="text-muted"
-                            attrs="{'invisible': [('ocr_google_enabled', '!=', False)]}"
+                            attrs="{'invisible': [('ocr_google_enabled', '!=', 'no_send')]}"
                         >
                             Enable Google OCR for invoices
                         </div>
+
+                        <div class="o_light_label">
+                            <field
+                                name="ocr_google_enabled"
+                                widget="radio"
+                                required="1"
+                            />
+                        </div>
                         <div
                             class="text-muted"
-                            attrs="{'invisible': [('ocr_google_enabled', '=', False)]}"
+                            attrs="{'invisible': [('ocr_google_enabled', '=', 'no_send')]}"
                         >
                             Access Google Cloud console, create an App, enable Google Document AI API and fill the following data
                         </div>
                         <div
                             class="content-group"
-                            attrs="{'invisible': [('ocr_google_enabled', '=', False)]}"
+                            attrs="{'invisible': [('ocr_google_enabled', '=', 'no_send')]}"
                         >
                             <div class="row mt16">
                                 <label
@@ -54,7 +60,7 @@
                                 />
                                 <field
                                     name="ocr_google_project"
-                                    attrs="{'required': [('ocr_google_enabled', '=', True)]}"
+                                    attrs="{'required': [('ocr_google_enabled', '!=', 'no_send')]}"
                                 />
                             </div>
                             <div class="row mt16">
@@ -65,7 +71,7 @@
                                 />
                                 <field
                                     name="ocr_google_location"
-                                    attrs="{'required': [('ocr_google_enabled', '=', True)]}"
+                                    attrs="{'required': [('ocr_google_enabled', '!=', 'no_send')]}"
                                 />
                             </div>
                             <div class="row mt16">
@@ -76,7 +82,7 @@
                                 />
                                 <field
                                     name="ocr_google_processor"
-                                    attrs="{'required': [('ocr_google_enabled', '=', True)]}"
+                                    attrs="{'required': [('ocr_google_enabled', '!=', 'no_send')]}"
                                 />
                             </div>
                             <div class="row mt16">
@@ -88,7 +94,7 @@
                                 <field
                                     name="ocr_google_authentication"
                                     filename="ocr_google_authentication_name"
-                                    attrs="{'required': [('ocr_google_enabled', '=', True)]}"
+                                    attrs="{'required': [('ocr_google_enabled', '!=', 'no_send')]}"
                                 />
                                 <field
                                     name="ocr_google_authentication_name"
